+++ server/src/main/java/com/accolite/server/readers/KeyResultExcelReader.java
@@ -32,7 +32,7 @@ public class KeyResultExcelReader {
                 keyResult.setDescription(row.getCell(4).getStringCellValue());
                 keyResult.setWeight((int) (row.getCell(5).getNumericCellValue()));
                 keyResult.setPeriod(row.getCell(6).getStringCellValue());
-                keyResult.setWindowId(row.getCell(7) != null && row.getCell(7).getCellType() == CellType.NUMERIC ? (int) row.getCell(7).getNumericCellValue() : null);
+                keyResult.setWindowId(row.getCell(7) != null && row.getCell(7).getCellType() == CellType.NUMERIC ? (long) row.getCell(7).getNumericCellValue() : null);
                 keyResult.setRating(row.getCell(8) != null && row.getCell(8).getCellType() == CellType.NUMERIC ? (int) row.getCell(8).getNumericCellValue() : null);
                 keyResults.add(keyResult);
             }
